Update and rename addons/autoLock/autoLock.lua to addons/autolock/autolock.lua
